final change

+++ store/models.py
@@ -56,7 +56,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name=models.CharField(max_length=200,null=True)
-    price=models.IntegerField()
+    price=models.DecimalField(max_digits=7,decimal_places=2)
     digital=models.BooleanField(default=True,null=True,blank=False)
     category=models.ForeignKey(Category,on_delete=models.CASCADE,null=True,blank=True)
     image=models.ImageField(null=True,blank=True)
